+++ tom_d4rt/lib/src/stdlib/typed_data.dart
@@ -1,4 +1,5 @@
 import 'package:tom_d4rt/src/environment.dart';
+import 'typed_data/bytes_builder.dart';
 import 'typed_data/endian.dart';
 import 'typed_data/byte_buffer.dart';
 import 'typed_data/uint8_list.dart';
@@ -16,6 +17,7 @@ import 'typed_data/float64_list.dart';
 
 class TypedDataStdlib {
   static void register(Environment environment) {
+    environment.defineBridge(BytesBuilderTypedData.definition);
     environment.defineBridge(EndianTypedData.definition);
     environment.defineBridge(ByteBufferTypedData.definition);
     environment.defineBridge(Uint8ListTypedData.definition);